+++ lib/ransack/configuration.rb
@@ -34,7 +34,9 @@ module Ransack
       :down_arrow => '&#9650;'.freeze,
       :default_arrow => nil,
       :remove_association_no_negative_assoc => true,
-      :sanitize_scope_args => true
+      :sanitize_scope_args => true,
+      :postgres_fields_sort_option => nil,
+      :strip_whitespace => true
     }
 
     def configure
@@ -142,6 +144,21 @@ module Ransack
       self.options[:sanitize_scope_args] = boolean
     end
 
+    # The `NULLS FIRST` and `NULLS LAST` options can be used to determine
+    # whether nulls appear before or after non-null values in the sort ordering.
+    #
+    # User may want to configure it like this:
+    #
+    # Ransack.configure do |c|
+    #   c.postgres_fields_sort_option = :nulls_first # or e.g. :nulls_always_last
+    # end
+    #
+    # See this feature: https://www.postgresql.org/docs/13/queries-order.html
+    #
+    def postgres_fields_sort_option=(setting)
+      self.options[:postgres_fields_sort_option] = setting
+    end
+
     # By default, Ransack displays sort order indicator arrows in sort links.
     # The default may be globally overridden in an initializer file like
     # `config/initializers/ransack.rb` as follows:
@@ -155,15 +172,19 @@ module Ransack
       self.options[:hide_sort_order_indicators] = boolean
     end
 
-    # By default, associations are removed from join dependencies when using negative predicates.
-    # The default may be globally overridden in an initializer file like
+    # By default, Ransack displays strips all whitespace when searching for a string.
+    # The default may be globally changed in an initializer file like
     # `config/initializers/ransack.rb` as follows:
     #
     # Ransack.configure do |config|
-    #   # Do not remove association from being joined when using negative predicates
-    #   config.remove_association_no_negative_assoc = false
+    #   # Enable whitespace stripping for string searches
+    #   config.strip_whitespace = true
     # end
     #
+    def strip_whitespace=(boolean)
+      self.options[:strip_whitespace] = boolean
+    end
+
     def remove_association_no_negative_assoc=(boolean)
       self.options[:remove_association_no_negative_assoc] = boolean
     end